requirements.txt
Flask==3.1.0
pillow==11.1.0
PyAutoGUI==0.9.54
pyinstaller==6.12.0
qrcode==8.0
waitress==3.0.2